Replace current one shot segmentation with two-steps (localization + segmentation), Include checkpoints corresponding to networks trained with WithoutEyesTrain dataset (biggest dataset)
