+++ docker/bidsapp/environment.yml
@@ -6,7 +6,7 @@ channels:
 - anaconda
 
 dependencies:
-- python=3.7
+- python=3.7.3
 - pip=19.0.3
 - nibabel=3.0.1
 - nipype=1.5.0
